Refuse a position this client cannot place or vouch for
The map lane read latitude and longitude off the wire and drew them. Three
refusals the browser makes were missing, and a comment here claimed one of
them was already being made.

Role. The browser settles which lane a group belongs to in its decoder, which
drops a group stamped with another role before any consumer is offered the
sample. This client decodes the sample without that step and hands the raw
wire type straight to the map, so nothing stood between a truth-stamped
position and the mark that says the estimator solved it — the substitution the
roles exist to stop. The fix and the truth lane are now gated on the role each
requires.

Datum. The schema says an unknown horizontal datum is refused at the receiver
and never guessed. Two datums put the same degrees a couple of metres apart
and which was meant is not recoverable afterwards, so a position stating none
is not placed, nor one naming a datum this build does not know, nor a
realization-bearing datum that does not say which realization.

Range. A latitude past the pole, a longitude the producer did not normalize,
or a coordinate that is not a number were all drawn. The browser refuses each,
and wrapping a longitude here rather than refusing it would draw the vehicle a
full turn of the Earth from where the other client draws nothing.

What is still not mirrored is stated in the code: the browser drops a fix
whose VERTICAL datum lacks the identity it requires, taking the position with
it, so a sample malformed only in its height draws a mark here and none there.
Closing that means consuming the geo contract rather than keeping a second
copy of its rules in this file.

The test module passed the size limit with the new cases, so it splits by what
each half asks: which lane the mark is read from, and what the mark points at.
